Rename forms/pattern.ts to forms/components.ts
