+++ packages/forms/src/index.ts
@@ -10,10 +10,10 @@ import {
 } from './element';
 
 export * from './builder';
+export * from './components';
 export * from './config';
 export * from './documents';
 export * from './element';
-export * from './pattern';
 export * from './response';
 export * from './session';
