+++ packages/forms/src/patterns/paragraph.ts
@@ -1,7 +1,7 @@
 import * as z from 'zod';
 
 import { type Pattern, type PatternConfig } from '../element';
-import { type PatternProps, type ParagraphPattern } from '../pattern';
+import { type PatternProps, type ParagraphPattern } from '../components';
 import { safeZodParse } from '../util/zod';
 
 const configSchema = z.object({
